Fixed the issue #131.
Fixed the issue that one more asterisk is added when starting a new line from a block comments.

Solution:
I modified the regular expression for the case that an asterisk needs to be added. The new expression will check that if a string is ended by an slash(/), a new asterisk will be added only when the string is not ended by an slash(/).