Remove hardcoded tab (`\t`) indentation
When pressing tab inside a comment `DocBlockrAtom.prototype.indent_command` is executed. This will copy the indention of the the line above or just do a normal indentation if the line above is not indented.
The normal indentation shouldn't be just a tab, rather fall back to the default behavior.

This will fix #163