Add C++ tests; no args is parsed as null
The C++ tests were largely based on the Java specs because they're
similar languages and have similar classes/primitive types. Added a
test specifically for template types.

For the C++ parser, if there were no args, it was apparently being
parsed as empty strings, which does not seem to be the protocol for the
other parsers, so changed it to evaluate to null instead.